pyproject.toml
[tool.poetry]
name = "myboost"
version = "0.1.0"
description = ""
authors = ["Tertius Olivier <tertius.olivier@cloudsmiths.ai>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
django-types = "^0.20.0"
django-stubs-ext = "^5.2.0"
djangorestframework = "^3.16.0"
django-environ = "^0.12.0"
django-mongodb-backend = "^5.2.0b0"
psycopg2 = "^2.9.10"
django-summernote = "^0.8.20.0"
django-daisy = "^1.0.23"
firebase-admin = "^6.8.0"
gunicorn = "^23.0.0"
pillow = "^11.2.1"
django-cors-headers = "^4.7.0"
django-cache-cleaner = "^0.2.0"
django = "^5.2.6"
django-google-maps = "^0.13.0"
pillow-avif-plugin = "^1.5.2"
django-redis = "^6.0.0"
redis = "^5.0.0"
pymemcache = "^4.0.0"
cryptography = "^46.0.3"
djangorestframework-simplejwt = "^5.5.1"
black = "^25.11.0"
django-vite = "^3.1.0"
drf-yasg = "^1.21.11"
xmltodict = "^1.0.2"
makefile = "^1.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
